[change]separate seed file
