+++ backend/db/seeds/direct_message_groups.rb
@@ -1,0 +1,7 @@
+ActiveRecord::Base.transaction do
+  test_user = User.first
+  User.where.not(id: test_user.id).limit(10).each do |item|
+    p "create dm group by_user_id:#{test_user.id} to_user_id:#{item.id}"
+    test_user.by_users.create!(to_user_id: item.id)
+  end
+end
